refactor(backend): extract sleep backends into separate modules
