+++ backend/app/services/power/sleep.py
@@ -83,195 +83,9 @@ class SleepBackend(abc.ABC):
         """Check if a system tool is available (hdparm, rtcwake, systemctl)."""
 
 
-class DevSleepBackend(SleepBackend):
-    """Development backend that simulates all sleep operations in-memory."""
-
-    def __init__(self) -> None:
-        self._spun_down: set[str] = set()
-        self._suspended = False
-
-    async def spindown_disks(self, devices: list[str]) -> list[str]:
-        self._spun_down.update(devices)
-        logger.info("[DEV] Simulated disk spindown: %s", devices)
-        return devices
-
-    async def spinup_disks(self, devices: list[str]) -> list[str]:
-        self._spun_down.discard(*devices) if len(devices) == 1 else self._spun_down.difference_update(devices)
-        logger.info("[DEV] Simulated disk spinup: %s", devices)
-        return devices
-
-    async def suspend_system(self) -> bool:
-        self._suspended = True
-        logger.info("[DEV] Simulated system suspend")
-        # Simulate waking up after a short delay
-        await asyncio.sleep(2)
-        self._suspended = False
-        logger.info("[DEV] Simulated system wake from suspend")
-        return True
-
-    async def schedule_rtc_wake(self, wake_at: datetime) -> bool:
-        logger.info("[DEV] Simulated RTC wake scheduled for %s", wake_at.isoformat())
-        return True
-
-    async def send_wol_packet(self, mac_address: str, broadcast_address: str) -> bool:
-        logger.info("[DEV] Simulated WoL packet sent to %s via %s", mac_address, broadcast_address)
-        return True
-
-    async def get_wol_capability(self) -> list[str]:
-        return ["eth0"]
-
-    async def get_data_disk_devices(self) -> list[str]:
-        return ["/dev/sda", "/dev/sdb"]
-
-    async def check_tool_available(self, tool: str) -> bool:
-        return True
-
-
-class LinuxSleepBackend(SleepBackend):
-    """Linux production backend using real system commands."""
-
-    async def _run_cmd(self, cmd: list[str], timeout: int = 10) -> tuple[bool, str]:
-        """Run a subprocess command safely with list args."""
-        import subprocess
-        try:
-            result = subprocess.run(
-                cmd,
-                capture_output=True,
-                text=True,
-                timeout=timeout,
-            )
-            return result.returncode == 0, result.stdout + result.stderr
-        except subprocess.TimeoutExpired:
-            logger.warning("Command timed out: %s", cmd)
-            return False, "timeout"
-        except Exception as e:
-            logger.warning("Command failed: %s — %s", cmd, e)
-            return False, str(e)
-
-    async def spindown_disks(self, devices: list[str]) -> list[str]:
-        spun_down = []
-        for device in devices:
-            ok, output = await self._run_cmd(["hdparm", "-y", device])
-            if ok:
-                spun_down.append(device)
-                logger.info("Disk spindown: %s", device)
-            else:
-                logger.warning("Failed to spin down %s: %s", device, output)
-        return spun_down
-
-    async def spinup_disks(self, devices: list[str]) -> list[str]:
-        spun_up = []
-        for device in devices:
-            # Reading from disk triggers spinup
-            ok, output = await self._run_cmd(["hdparm", "-C", device])
-            if ok:
-                spun_up.append(device)
-                logger.info("Disk spinup: %s", device)
-            else:
-                logger.warning("Failed to spin up %s: %s", device, output)
-        return spun_up
-
-    async def suspend_system(self) -> bool:
-        ok, output = await self._run_cmd(["sudo", "systemctl", "suspend"], timeout=30)
-        if not ok:
-            logger.error("System suspend failed: %s", output)
-        return ok
-
-    async def schedule_rtc_wake(self, wake_at: datetime) -> bool:
-        timestamp = str(int(wake_at.timestamp()))
-        ok, output = await self._run_cmd(["sudo", "rtcwake", "-m", "no", "-l", "-t", timestamp])
-        if not ok:
-            logger.error("RTC wake schedule failed: %s", output)
-        return ok
-
-    async def send_wol_packet(self, mac_address: str, broadcast_address: str) -> bool:
-        """Send WoL magic packet via raw socket (no subprocess needed)."""
-        try:
-            mac_bytes = bytes.fromhex(mac_address.replace(":", "").replace("-", ""))
-            magic = b'\xff' * 6 + mac_bytes * 16
-            sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-            sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
-            sock.sendto(magic, (broadcast_address, 9))
-            sock.close()
-            logger.info("WoL packet sent to %s via %s", mac_address, broadcast_address)
-            return True
-        except Exception as e:
-            logger.error("Failed to send WoL packet: %s", e)
-            return False
-
-    async def get_wol_capability(self) -> list[str]:
-        """Check which interfaces support WoL via ethtool."""
-        import subprocess
-        interfaces = []
-        # Get network interface names
-        try:
-            result = subprocess.run(
-                ["ip", "-o", "link", "show"],
-                capture_output=True, text=True, timeout=5,
-            )
-            if result.returncode == 0:
-                for line in result.stdout.splitlines():
-                    parts = line.split(":")
-                    if len(parts) >= 2:
-                        iface = parts[1].strip()
-                        if iface != "lo":
-                            # Check WoL support
-                            wol_result = subprocess.run(
-                                ["ethtool", iface],
-                                capture_output=True, text=True, timeout=5,
-                            )
-                            if wol_result.returncode == 0 and "Wake-on:" in wol_result.stdout:
-                                for wol_line in wol_result.stdout.splitlines():
-                                    if "Wake-on:" in wol_line and "d" not in wol_line.split(":")[-1]:
-                                        interfaces.append(iface)
-                                        break
-        except Exception as e:
-            logger.warning("Failed to check WoL capability: %s", e)
-        return interfaces
-
-    async def get_data_disk_devices(self) -> list[str]:
-        """Get data disk devices excluding the OS disk (same pattern as RAID backend)."""
-        import subprocess
-        try:
-            result = subprocess.run(
-                ["lsblk", "-J", "-o", "NAME,TYPE,MOUNTPOINTS"],
-                capture_output=True, text=True, timeout=10,
-            )
-            if result.returncode != 0:
-                return []
-
-            data = json.loads(result.stdout)
-            os_disk_name: Optional[str] = None
-            all_disks: list[str] = []
-
-            def _check_mountpoints(device: dict) -> bool:
-                """Recursively check if device or children have root mount."""
-                mps = device.get("mountpoints", [])
-                if mps and "/" in mps:
-                    return True
-                for child in device.get("children", []):
-                    if _check_mountpoints(child):
-                        return True
-                return False
-
-            for device in data.get("blockdevices", []):
-                if device.get("type") == "disk":
-                    name = device["name"]
-                    if _check_mountpoints(device):
-                        os_disk_name = name
-                    else:
-                        all_disks.append(f"/dev/{name}")
-
-            if os_disk_name:
-                logger.debug("OS disk detected: %s (excluded from spindown)", os_disk_name)
-
-            return all_disks
-        except Exception as e:
-            logger.warning("Failed to enumerate data disks: %s", e)
-            return []
-
-    async def check_tool_available(self, tool: str) -> bool:
-        return shutil.which(tool) is not None
+# Re-export backend classes for backward compatibility
+from app.services.power.sleep_backend_dev import DevSleepBackend  # noqa: E402, F401
+from app.services.power.sleep_backend_linux import LinuxSleepBackend  # noqa: E402, F401
 
 
 # ---------------------------------------------------------------------------
